+++ src/types.ts
@@ -16,18 +16,16 @@ export type Currency = string
 export type FX_Rate = number
 export type Fee = number
 
-export type Transaction = [
-    Date,
-    Ticker | null,
-    Type,
-    Quantity | null,
-    PricePerShare | null,
-    Amount,
-    Currency,
-    FX_Rate,
-    Fee | null
-]
+export type Transaction = {
+    date: Date
+    ticker: Ticker | null
+    type: Type
+    quantity: Quantity | null
+    pricePerShare: PricePerShare | null
+    amount: Amount
+    currency: Currency
+    fxRate: FX_Rate
+    fee: Fee | null
+}
 
 export type GroupedTypes = Record<Type, Transaction[]>
-
-export type SellsSummaryRow = {}