+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Sells summary (FIFO)
 
+### Fixed
+
+-   Transaction type
+
 ## [v0.1.0] - 2022-03-13
 
 ### Added